refactor(schema): Adopt RustVersion
