+++ src/cargo/core/package.rs
@@ -32,8 +32,8 @@ use crate::util::network::http::HttpTimeout;
 use crate::util::network::retry::{Retry, RetryResult};
 use crate::util::network::sleep::SleepTracker;
 use crate::util::toml::prepare_for_publish;
-use crate::util::RustVersion;
 use crate::util::{self, internal, Config, Progress, ProgressStyle};
+use crate::util_schemas::manifest::RustVersion;
 
 pub const MANIFEST_PREAMBLE: &str = "\
 # THIS FILE IS AUTOMATICALLY GENERATED BY CARGO
