+++ src/cargo/core/workspace.rs
@@ -25,8 +25,8 @@ use crate::util::edit_distance;
 use crate::util::errors::{CargoResult, ManifestError};
 use crate::util::interning::InternedString;
 use crate::util::toml::{read_manifest, InheritableFields};
-use crate::util::RustVersion;
 use crate::util::{config::ConfigRelativePath, Config, Filesystem, IntoUrl};
+use crate::util_schemas::manifest::RustVersion;
 use crate::util_schemas::manifest::{TomlDependency, TomlProfiles};
 use cargo_util::paths;
 use cargo_util::paths::normalize_path;
