+++ src/cargo/ops/resolve.rs
@@ -71,8 +71,8 @@ use crate::ops;
 use crate::sources::PathSource;
 use crate::util::cache_lock::CacheLockMode;
 use crate::util::errors::CargoResult;
-use crate::util::RustVersion;
 use crate::util::{profile, CanonicalUrl};
+use crate::util_schemas::manifest::RustVersion;
 use anyhow::Context as _;
 use std::collections::{HashMap, HashSet};
 use tracing::{debug, trace};
