+++ src/cargo/sources/registry/index.rs
@@ -92,7 +92,8 @@ use crate::sources::registry::{LoadResponse, RegistryData};
 use crate::util::cache_lock::CacheLockMode;
 use crate::util::interning::InternedString;
 use crate::util::IntoUrl;
-use crate::util::{internal, CargoResult, Config, Filesystem, OptVersionReq, RustVersion};
+use crate::util::{internal, CargoResult, Config, Filesystem, OptVersionReq};
+use crate::util_schemas::manifest::RustVersion;
 use anyhow::bail;
 use cargo_util::{paths, registry::make_dep_path};
 use semver::Version;
