+++ src/cargo/util/semver_ext.rs
@@ -1,9 +1,7 @@
 use std::fmt::{self, Display};
 
 use semver::{Op, Version, VersionReq};
-use serde_untagged::UntaggedEnumVisitor;
 
-use crate::util_semver::PartialVersion;
 use crate::util_semver::VersionExt as _;
 
 #[derive(PartialEq, Eq, Hash, Clone, Debug)]
@@ -123,50 +121,5 @@ impl Display for OptVersionReq {
 impl From<VersionReq> for OptVersionReq {
     fn from(req: VersionReq) -> Self {
         OptVersionReq::Req(req)
-    }
-}
-
-#[derive(PartialEq, Eq, PartialOrd, Ord, Hash, Clone, Debug, serde::Serialize)]
-#[serde(transparent)]
-pub struct RustVersion(PartialVersion);
-
-impl std::ops::Deref for RustVersion {
-    type Target = PartialVersion;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl std::str::FromStr for RustVersion {
-    type Err = anyhow::Error;
-
-    fn from_str(value: &str) -> Result<Self, Self::Err> {
-        let partial = value.parse::<PartialVersion>()?;
-        if partial.pre.is_some() {
-            anyhow::bail!("unexpected prerelease field, expected a version like \"1.32\"")
-        }
-        if partial.build.is_some() {
-            anyhow::bail!("unexpected prerelease field, expected a version like \"1.32\"")
-        }
-        Ok(Self(partial))
-    }
-}
-
-impl<'de> serde::Deserialize<'de> for RustVersion {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        UntaggedEnumVisitor::new()
-            .expecting("SemVer version")
-            .string(|value| value.parse().map_err(serde::de::Error::custom))
-            .deserialize(deserializer)
-    }
-}
-
-impl Display for RustVersion {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.fmt(f)
     }
 }
